lifesync-ai/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
websockets==12.0
langchain==0.2.1
langchain-openai==0.1.8
langgraph==0.1.1
openai==1.30.1
chromadb==0.5.0
sentence-transformers==3.0.0
stable-baselines3==2.3.2
gymnasium==0.29.1
optuna==3.6.1
open-clip-torch==2.24.0
mediapipe==0.10.14
ultralytics==8.2.18
boto3==1.34.0
openai-whisper==20231117
gTTS==2.5.1
python-dotenv==1.0.1
pydantic==2.7.1
psycopg2-binary==2.9.9
pg8000==1.31.5
httpx==0.27.0
pytest==8.2.0
pytest-asyncio==0.23.7
